+++ __tests__/natural.flash.test.js
@@ -1,31 +1,31 @@
 var app = require('../tests')
 var supertest = require('supertest')
-var natural = require('../index')
-var NaturalFlashTest = require('./data/natural.flash')
+var request = supertest(app)
+var natural = require('../lib_clone/index')
+var naturalFlashTest = require('./data/natural_flash')
 
 describe('group all test flash message', () => {
   it('multiple test flash message send 50 error messages', async (done) => {
-    await natural.routeMiddleware(app, [NaturalFlashTest])
 
     const errorMessages = []
+    await natural.pluginMiddleware(app, [natural.flashExpress()])
+    await natural.routeMiddleware(app, [naturalFlashTest()])
+    const res = await request.get('/tests5');
+    errorMessages.push(res.body.errors);
+    expect(typeof res.body.errors).toBeDefined()
+    expect(typeof res.body.errors).toStrictEqual('object')
 
-    const errorsData = await global.fs.readFileSync(global.path.join(__dirname, '../small.dummy.json'),
-      { encoding: 'utf-8' })
-    errorMessages.push(errorsData)
-    const newErrorMessages = JSON.parse(errorsData)
-
-    // POST METHOD
-    supertest(app)
-      .post('/flash')
-      .send({ errors: newErrorMessages })
-      .set('Accept', 'application/json')
-      .expect('Content-Type', /json/)
-      .expect(200)
-      .end(function (err, res) {
-        if (err) return done(err)
-        done()
-        expect(typeof res.body.data).toBeDefined()
-        expect(typeof res.body.data).toStrictEqual('object')
-      })
-  })
+         supertest(app)
+          .post('/tests6')
+          .send({errors: errorMessages})
+          .set('Accept', 'application/json')
+          .expect('Content-Type', /json/)
+          .expect(200)
+          .end((err, res) => {
+             if(err) new Error(err)
+             done()
+             expect(typeof res.body.errors).toBeDefined()
+             expect(typeof res.body.errors).toStrictEqual('object')
+          })
+    })
 })