+++ example/express/index.js
@@ -1,17 +1,17 @@
-const natural = require("../../index");
-
-// injected for global module
+// register all module
+const natural = require("../../lib_clone/index");
 natural.globalModule(
-["express", "http", "mongoose", "bodyParser", "cookieParser", "logger"],
-["express", "http", "mongoose", "body-parser", "cookie-parser", "morgan"]);
+["express", "http", "mongoose", "bodyParser", "cookieParser", "logger", "path"],
+["express", "http", "mongoose", "body-parser", "cookie-parser", "morgan", "path"]);
 
-// init module
+// init all module
 const app = express();
 const router = express.Router();
 const server = http.createServer(app);
 
-// init user route
-const indexRoute = require("./routes/user.route")
+// init all route
+const indexRoute = require("./routes/index.route");
+const userRoute = require("./routes/user.route")
 
 // setup global promise
 mongoose.Promise = global.Promise;
@@ -21,17 +21,18 @@ mongoose.connect("mongodb://localhost:27017/demo",
 {useNewUrlParser: true, useUnifiedTopology: true,
 useCreateIndex: true, useFindAndModify: false})
 .then(() => console.log("database connected successfully"))
-.catch(() => console.log("database connection failed"));;
+.catch(() => console.log("database connection failed"));
 
-//  register all plugin middleware with chaining middleware
+//  register all plugin middleware
 natural.pluginMiddleware(app, [
     bodyParser.urlencoded({ extended: false }),
     bodyParser.json(),
     cookieParser(),
+    natural.flashExpress(),
     logger("dev")
 ]);
 
-//  register all plugin middleware with chaining middleware
+//  register all plugin middleware using chaining middleware
 // natural.pluginMiddlewareAsync(app, [
 //     bodyParser.urlencoded({ extended: false }),
 //     bodyParser.json(),
@@ -39,12 +40,14 @@ natural.pluginMiddleware(app, [
 // ])
 //.use(logger("dev"));
 
-// register to route middleware
-natural.routeMiddleware(app, [indexRoute]);
+// register all route middleware
+natural.routeMiddleware(app, [indexRoute, userRoute]);
 
-// set configs
+// set configs and template engine
+app.set("views", path.join(__dirname, "/views/"));
+app.set("view engine", "ejs");
 app.set("env", process.env.PORT || 3001);
 app.set("x-powered-by", false);
 
-// listeing server
+// listening server
 server.listen(app.get("env"), () => console.log("server is running"));